requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.10.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
filelock==3.19.1
frozenlist==1.5.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.32.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.14.0
h2==4.3.0
hf-xet==1.1.8
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-huggingface==0.3.1
langchain-openai==0.3.4
langchain-pinecone==0.2.2
langchain-tests==0.3.11
langchain-text-splitters==0.3.9
langchainhub==0.1.21
langsmith==0.3.45
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.26.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.61.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pinecone==5.4.2
pinecone-plugin-assistant==1.7.0
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
postgrest==1.1.1
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.3.0
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
realtime==2.7.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
storage3==0.12.1
streamlit==1.42.0
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
sympy==1.14.0
syrupy==4.8.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.21.4
toml==0.10.2
torch==2.8.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.55.4
triton==3.4.0
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.1
uritemplate==4.2.0
urllib3==2.3.0
watchdog==6.0.0
websockets==15.0.1
yarl==1.18.3
zstandard==0.23.0
